requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.8.5
requests==2.32.5
